A simple implementation of opposition control, and additional analysis routines
pipeSimpleResolventBC:  A simple way to implement opposition control
within the resolvent analysis framework.  This simply sets the
wall-normal velocity at the bed to be equal and opposite to the wall
normal velocity at a given elevation.

Most of the other functions added have to do with analysis of the
results.
- fourier2physical: converts the helical propagating waves to physical
space
- uGradFourier: calculates the velocity gradient tensor in fourier space
- pipeGreenFn: calculates the Green function for pressure
- pipeGreenU0: calculates the Green function for the mean velocity
- multiModeSwirl: calculates the swirl field for a combination of waves